+++ examples/incubator/bitcoin-payment-request-example-application/src/main/java/org/tbk/bitcoin/example/payreq/bitcoin/block/BitcoinBlocks.java
@@ -1,6 +1,6 @@
 package org.tbk.bitcoin.example.payreq.bitcoin.block;
 
-import org.jmolecules.spring.AssociationResolver;
+import org.jmolecules.ddd.integration.AssociationResolver;
 import org.springframework.data.jpa.repository.JpaSpecificationExecutor;
 import org.springframework.data.repository.CrudRepository;
 import org.springframework.data.repository.PagingAndSortingRepository;